pyproject.toml
[build-system]
requires = ["setuptools >= 61.0", "wheel", "uv >= 0.1.17"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-python-streamable-e2e-test-template"
version = "0.1.0"
description = "A template for creating Dockerized Python FastMCP (Model Context Protocol) servers with E2E testing capabilities. Based on the mcp-sdk."
keywords = ["MCP", "FastMCP", "template", "docker", "python"]
authors = [
    { name = "akitana-airtanker", email = "akitana_airtanker-github@yahoo.co.jp" },
    { name = "sebastus", email = "divingclimber@gmail.com" },
]
license = { text = "MIT License" }
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    "mcp[cli] @ git+https://github.com/modelcontextprotocol/python-sdk.git",
    "azure-monitor-opentelemetry",
    "python-dotenv",
]

[project.urls]
Homepage = "https://github.com/sebastus/mcp-python-streamable-e2e-test-template"
Repository = "https://github.com/sebastus/mcp-python-streamable-e2e-test-template"

[project.optional-dependencies]
test = ["pytest", "pytest-asyncio", "uv"]

[project.scripts]
mcp-server-demo = "mcp_server.server:main"
mcp-client-demo = "mcp_python_streamable_e2e_test_template.client:main"

[tool.setuptools]
package-dir = { "" = "src" }
packages = ["mcp_server"]

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
